Fixing a rounding error in the timestamps.  You have to use double and not float.
git-svn-id: svn+ssh://svn.code.sf.net/p/swistrack/code@938 18c3c8b4-d610-0410-8eed-94a590f67c64

+++ Code/core/DataAssociationClassifier.h
@@ -14,7 +14,7 @@ using namespace std;
 class DataAssociationClassifier {
 
  public: 
-  typedef map<string, float> Example;
+  typedef map<string, double> Example;
   typedef vector<Example> ExampleTable;
